+++ util/misc_test.go
@@ -81,49 +81,6 @@ func (s *testMiscSuite) TestRunWithRetry(c *C) {
 	c.Assert(cnt, Equals, 1)
 }
 
-func (s *testMiscSuite) TestCompatibleParseGCTime(c *C) {
-	values := []string{
-		"20181218-19:53:37 +0800 CST",
-		"20181218-19:53:37 +0800 MST",
-		"20181218-19:53:37 +0800 FOO",
-		"20181218-19:53:37 +0800 +08",
-		"20181218-19:53:37 +0800",
-		"20181218-19:53:37 +0800 ",
-		"20181218-11:53:37 +0000",
-	}
-
-	invalidValues := []string{
-		"",
-		" ",
-		"foo",
-		"20181218-11:53:37",
-		"20181218-19:53:37 +0800CST",
-		"20181218-19:53:37 +0800 FOO BAR",
-		"20181218-19:53:37 +0800FOOOOOOO BAR",
-		"20181218-19:53:37 ",
-	}
-
-	expectedTime := time.Date(2018, 12, 18, 11, 53, 37, 0, time.UTC)
-	expectedTimeFormatted := "20181218-19:53:37 +0800"
-
-	beijing, err := time.LoadLocation("Asia/Shanghai")
-	c.Assert(err, IsNil)
-
-	for _, value := range values {
-		t, err := CompatibleParseGCTime(value)
-		c.Assert(err, IsNil)
-		c.Assert(t.Equal(expectedTime), Equals, true)
-
-		formatted := t.In(beijing).Format(GCTimeFormat)
-		c.Assert(formatted, Equals, expectedTimeFormatted)
-	}
-
-	for _, value := range invalidValues {
-		_, err := CompatibleParseGCTime(value)
-		c.Assert(err, NotNil)
-	}
-}
-
 func (s *testMiscSuite) TestX509NameParseMatch(c *C) {
 	check := pkix.Name{
 		Names: []pkix.AttributeTypeAndValue{
